Double for Number instead of Float.
Even as I was doing this port I thought I might need to do this, and it is indeed more compatible
with the Actionscript implementation. The immediate need comes out of streaming Dates as timestamps.
The current number of seconds since the epoch is already greater than a 32-bit int can hold, so
neither int nor float is good enough. Java uses longs for timestamps, but double will be good enough
basically forever, and it matches what the AS implementation does (mainly because the only way to
get a 64-bit number in AS is via the floating point Number).

+++ runtimes/java/src/main/java/com/microtome/core/Annotation.java
@@ -3,6 +3,6 @@ package com.microtome.core;
 public interface Annotation {
     boolean boolValue (boolean defaultVal);
     int intValue (int defaultVal);
-    float numberValue (float defaultVal);
+    double numberValue (double defaultVal);
     String stringValue (String defaultVal);
 }
