+++ runtimes/java/src/main/java/com/microtome/marshaller/NumberMarshaller.java
@@ -9,32 +9,32 @@ import com.microtome.core.WritableObject;
 import com.microtome.error.MicrotomeError;
 import com.microtome.prop.Prop;
 
-public class NumberMarshaller extends PrimitiveMarshaller<Float> {
+public class NumberMarshaller extends PrimitiveMarshaller<Double> {
     public NumberMarshaller () {
-        super(Float.class);
+        super(Double.class);
     }
 
-    @Override public void validateProp (Prop<Float> prop) {
-        float min = prop.annotation(Defs.MIN_ANNOTATION).numberValue(Float.MIN_VALUE);
+    @Override public void validateProp (Prop<Double> prop) {
+        double min = prop.annotation(Defs.MIN_ANNOTATION).numberValue(Double.MIN_VALUE);
         if (prop.value() < min) {
             throw new MicrotomeError("Value too small [" + prop.value() + " < " + min + "]");
         }
-        float max = prop.annotation(Defs.MAX_ANNOATATION).numberValue(Float.MAX_VALUE);
+        double max = prop.annotation(Defs.MAX_ANNOATATION).numberValue(Double.MAX_VALUE);
         if (prop.value() > max) {
             throw new MicrotomeError("Value too large [" + prop.value() + " > " + max + "]");
         }
     }
 
-    @Override public Float readValue (MicrotomeMgr mgr, DataReader reader, String name,
+    @Override public Double readValue (MicrotomeMgr mgr, DataReader reader, String name,
             TypeInfo type) {
         return reader.requireNumber(name);
     }
 
-    @Override public Float readDefault (MicrotomeMgr mgr, TypeInfo type, Annotation anno) {
+    @Override public Double readDefault (MicrotomeMgr mgr, TypeInfo type, Annotation anno) {
         return anno.numberValue(0);
     }
 
-    @Override public void writeValue (MicrotomeMgr mgr, WritableObject writer, Float value,
+    @Override public void writeValue (MicrotomeMgr mgr, WritableObject writer, Double value,
             String name, TypeInfo type) {
         writer.writeNumber(name, value);
     }
